requirements.txt
Flask
tensorflow
pillow
cassandra-driver